backend/foodgram/requirements.txt
requests==2.26.0
django==2.2.19
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-filter==21.1
drf-extra-fields==3.4.0
djoser==2.1.0
Pillow==8.3.1
python-dotenv==0.20.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
weasyprint==52.5